requirements.txt
torch
torchvision
numpy
opencv-python
dlib
tqdm
scikit-learn
scipy
matplotlib
scikit-image
facenet-pytorch
Pillow